refactor(lilgit/model): refactored nodey into its own package for easier debugging
